Add half-step place ratings

+++ Wander/Services/Remote/SupabaseDTOs.swift
@@ -85,7 +85,7 @@ struct RemoteVisiblePlaceDTO: Codable, Equatable {
     let visibility: String
     let note: String?
     let ratingSignal: String?
-    let ratingScore: Int?
+    let ratingScore: Double?
     let recommendedScore: Double?
     let recommendedCount: Int?
     let categoryOverride: String?
